pyproject.toml
[project]
name = "agentmode"
version = "0.1.9"
description = "All-in-1 MCP server for developers"
readme = "README.md"
requires-python = ">=3.13"
keywords = ["mcp", "ai"]
dependencies = [
    "click>=8.1.8",
    "gradio>=5.25.2",
    "greenlet>=3.2.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "pandas>=2.2.3",
    "platformdirs>=4.3.7",
    "python-benedict[toml]>=0.34.1",
    "sqlalchemy>=2.0.40",
    "sqlglot[rs]>=26.16.0",
    "starlette>=0.46.2",
    "uvicorn>=0.34.1",
]

[project.scripts]
agentmode = "agentmode:cli"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "litellm>=1.67.4.post1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.6",
]

[project.urls]
"Homepage" = "https://www.agentmode.app"

[project.optional-dependencies]
database = [
    "aiomysql>=0.2.0",
    "aioodbc>=0.5.0",
    "aiotrino>=0.3.0",
    "asyncpg>=0.30.0",
    "clickhouse-sqlalchemy>=0.3.2",
    "oracledb>=3.1.0",
    "presto-python-client>=0.8.4",
    "pyathena[sqlalchemy]>=3.13.0",
    "pyhive>=0.7.0",
    "snowflake-sqlalchemy>=1.7.3",
    "sqlalchemy-bigquery>=1.13.0",
    "sqlalchemy-cockroachdb>=2.0.2",
    "sqlalchemy-hana>=3.0.2",
    "teradatasqlalchemy>=20.0.0.4",
]

[tool.hatch.build]
include = [
  "agentmode/**/*.py",  # Include all Python source files
  "agentmode/connectors.toml", 
  "agentmode/api/connectors/*.json", 
  "agentmode/resources/**/*", 
  "agentmode/resources/log_config.json"
]
